requirements.txt
gymnasium[atari]>=0.29.0
ale-py>=0.8.1
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0
pygame>=2.5.0
